+++ src/pages/taehoon/Login/LoginTaehoon.scss
@@ -1,8 +1,6 @@
-@import '../../../styles/common.scss';
 @import '../../../styles/variables.scss';
-@import '../../../styles/reset.scss';
 
-.login {
+.loginTaehoon {
   width: 25%;
   margin: 200px auto;
 
@@ -26,7 +24,7 @@
     width: 100%;
     padding-bottom: 10px;
 
-    #id {
+    .id {
       width: 100%;
       padding: 10px;
     }
@@ -36,7 +34,7 @@
     width: 100%;
     padding-bottom: 10px;
 
-    #password {
+    .password {
       width: 100%;
       padding: 10px;
     }
