+++ src/pages/taehoon/Main/MainTaehoon.scss
@@ -2,7 +2,7 @@
 @import '../../../styles/reset.scss';
 @import '../../../styles/variables.scss';
 
-.main {
+.mainTaehoon {
   width: 100%;
 
   .bar {
